examples/entity_extraction/pyproject.toml
[project]
name = "textual-haystack-entity-extraction-example"
version = "0.1.0"
description = "Example: PII entity extraction with Tonic Textual and Haystack"
requires-python = ">=3.10.0,<4.0.0"
dependencies = [
    "textual-haystack",
    "haystack-ai",
]

[tool.uv.sources]
textual-haystack = { path = "../.." }
